refactor(ProfileForm): Wydzielono logikę do custom hooków i podzielono komponent na mniejsze części zgodnie z zasadami SRP i DRY
