+++ src/components/Dashboard/SettingsSection/SettingsContent/ProfileForm/components/Avatar/Avatar.styles.ts
