+++ src/components/Dashboard/SettingsSection/SettingsContent/ProfileForm/components/UserInfoFields/UserInfoFields.component.tsx
@@ -1,9 +1,9 @@
 import { memo } from "react";
 import { FaUser, FaEnvelope } from "react-icons/fa";
-import { FormInput } from "../../../../../UI/Form/FormInput/FormInput.component";
+import { FormInput } from "../../../../../../UI/Form/FormInput/FormInput.component";
 import type { UseFormRegister, FieldErrors } from "react-hook-form";
-import type { UserProfile } from "../../../types/settings";
-import { styles } from "./UserInfoFields.styles";
+import type { UserProfile } from "../../../../types/settings";
+import { styles } from '../../style/ProfileForm.styles';
 
 type UserInfoFieldsProps = {
   register: UseFormRegister<UserProfile>;
